refactor(file): Move FileState to its own file.
